Add test helpers file
